Add files via upload
A separate class for a Button as there are a few of them. There is a constructor with button corners coordinates and a text inside of it. Also, there are methods to draw the button on a screen, to highlight it as the user hovers over it and to check if user's mouse position is on the button.